# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: General Supplies Procurement Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── INQUIRIES ───────────────────────────────────────────────────────────────
...
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List all quotations
      responses:
        "200":
          description: List of quotations with items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuotationWithItems"
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Create a new quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      summary: Get a quotation by ID (with items)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationWithItems"
        "404":
          description: Not found
    patch:
      operationId: updateQuotation
      tags: [quotations]
      summary: Update a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      summary: Delete a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotations/{id}/items:
    post:
      operationId: addQuotationItem
      tags: [quotations]
      summary: Add an item to a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationItem"
  /quotations/{id}/items/{itemId}:
    patch:
      operationId: updateQuotationItem
      tags: [quotations]
      summary: Update a quotation item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationItem"
    delete:
      operationId: deleteQuotationItem
      tags: [quotations]
      summary: Delete a quotation item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotations/{id}/approve:
    post:
      operationId: approveQuotation
      tags: [quotations]
      summary: Mark a quotation as approved by the customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
  /quotations/{id}/reject:
    post:
      operationId: rejectQuotation
      tags: [quotations]
      summary: Mark a quotation as rejected by the customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"

  # ─── CUSTOMER POs ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Supplier:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        category:
          type: string
        notes:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        category:
          type: string
        notes:
          type: string

    # ─── INQUIRY ───────────────────────────────────────────────────────────────
...
    Quotation:
      type: object
      required: [id, inquiryId, customerId, status, createdAt]
      properties:
        id:
          type: integer
        inquiryId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        quotationNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        totalAmount:
          type: ["number", "null"]
        validUntil:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    QuotationItem:
      type: object
      required: [id, quotationId, description, quantity, unitPrice]
      properties:
        id:
          type: integer
        quotationId:
          type: integer
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        description:
          type: string
        quantity:
          type: number
        unit:
          type: ["string", "null"]
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
    QuotationWithItems:
      type: object
      required: [id, inquiryId, customerId, status, createdAt, items]
      properties:
        id:
          type: integer
        inquiryId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        quotationNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        totalAmount:
          type: ["number", "null"]
        validUntil:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuotationItem"
    QuotationInput:
      type: object
      required: [inquiryId, customerId]
      properties:
        inquiryId:
          type: integer
        customerId:
          type: integer
        quotationNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        totalAmount:
          type: number
        validUntil:
          type: string
        notes:
          type: string
    QuotationUpdate:
      type: object
      properties:
        quotationNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        totalAmount:
          type: number
        validUntil:
          type: string
        notes:
          type: string
    QuotationItemInput:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        supplierId:
          type: integer
        description:
          type: string
          minLength: 1
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        notes:
          type: string
    QuotationItemUpdate:
      type: object
      properties:
        supplierId:
          type: integer
        description:
          type: string
          minLength: 1
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        notes:
          type: string

    # ─── CUSTOMER PO ───────────────────────────────────────────────────────────
...