# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * General Supplies Procurement Management System
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  /** @minLength 1 */
  name?: string;
  contactName?: string;
  email?: string;
  phone?: string;
  address?: string;
  category?: string;
  notes?: string;
}